feat: get prepare module ready for testing
