+++ src/superface/config/prepare-files.ts
@@ -1,7 +1,4 @@
 import {
-  assertMapDocumentNode,
-  assertProfileDocumentNode,
-  assertProviderJson,
   MapDocumentNode,
   ProfileDocumentNode,
   ProviderJson,
@@ -9,22 +6,18 @@ import {
 import {
   IFileSystem,
   NodeFileSystem,
-  Profile,
   profileAstId,
-  Provider,
   SuperJson,
 } from '@superfaceai/one-sdk';
-import { parseMap, parseProfile, Source } from '@superfaceai/parser';
 
 import {
   ComponentUndefinedError,
-  MapUndefinedError,
-  ProfileUndefinedError,
-  ProviderJsonUndefinedError,
   SuperJsonNotFoundError,
 } from '../../common/errors';
 import { SuperfaceTestConfigPayload } from '../../superface-test.interfaces';
 import { getSuperJson } from '../../superface-test.utils';
+import { getMapAst, getProfileAst } from './prepare-ast';
+import { getProviderJson } from './prepare-provider-json';
 
 // This deals only with files resolution, it should NOT be exported from directory.
 export async function prepareFiles(
@@ -74,83 +67,4 @@ export async function prepareFiles(
     providerJson,
     mapAst,
   };
-}
-
-async function getProfileAst(
-  profile: Profile | string,
-  superJson: SuperJson,
-  fileSystem: IFileSystem
-): Promise<ProfileDocumentNode> {
-  if (profile instanceof Profile) {
-    return profile.ast;
-  } else {
-    const profileSettings = superJson.normalized.profiles[profile];
-
-    if (!profileSettings || !('file' in profileSettings)) {
-      throw new ProfileUndefinedError(profile);
-    }
-
-    const profilePath = superJson.resolvePath(profileSettings.file);
-    const content = await fileSystem.readFile(profilePath);
-
-    if (content.isErr()) {
-      throw content.error;
-    }
-
-    if (profilePath.endsWith('.supr')) {
-      return parseProfile(new Source(content.value, profilePath));
-    } else {
-      return assertProfileDocumentNode(JSON.parse(content.value));
-    }
-  }
-}
-
-async function getMapAst(
-  profileId: string,
-  providerName: string,
-  superJson: SuperJson,
-  fileSystem: IFileSystem
-): Promise<MapDocumentNode> {
-  const profileProviderSettings =
-    superJson.normalized.profiles[profileId]?.providers[providerName];
-
-  if (!profileProviderSettings || !('file' in profileProviderSettings)) {
-    throw new MapUndefinedError(profileId, providerName);
-  }
-
-  const mapPath = superJson.resolvePath(profileProviderSettings.file);
-  const content = await fileSystem.readFile(mapPath);
-
-  if (content.isErr()) {
-    throw content.error;
-  }
-
-  if (mapPath.endsWith('.suma')) {
-    return parseMap(new Source(content.value, mapPath));
-  } else {
-    return assertMapDocumentNode(JSON.parse(content.value));
-  }
-}
-
-async function getProviderJson(
-  provider: Provider | string,
-  superJson: SuperJson,
-  fileSystem: IFileSystem
-): Promise<ProviderJson> {
-  const providerName =
-    provider instanceof Provider ? provider.configuration.name : provider;
-  const providerSettings = superJson.normalized.providers[providerName];
-
-  if (!providerSettings || !providerSettings.file) {
-    throw new ProviderJsonUndefinedError(providerName);
-  }
-
-  const providerPath = superJson.resolvePath(providerSettings.file);
-  const content = await fileSystem.readFile(providerPath);
-
-  if (content.isErr()) {
-    throw content.error;
-  }
-
-  return assertProviderJson(JSON.parse(content.value));
 }
