+++ src/superface/mock/boundProfileProvider.ts
@@ -6,37 +6,43 @@ import {
   ProfileParameterError,
   Result,
   SecurityConfiguration,
+  ServiceSelector,
 } from '@superfaceai/one-sdk';
 
 export const mockBoundProfileProvider = jest.fn<
   BoundProfileProvider,
   Parameters<
     (
-      configuration: {
-        services: IServiceSelector;
-        profileProviderSettings?: NormalizedProfileProviderSettings;
-        security: SecurityConfiguration[];
-        parameters?: Record<string, string>;
-      },
-
       performResult: Result<
         unknown,
         ProfileParameterError | MapInterpreterError
-      >
+      >,
+
+      configuration?: {
+        services?: IServiceSelector;
+        profileProviderSettings?: NormalizedProfileProviderSettings;
+        security?: SecurityConfiguration[];
+        parameters?: Record<string, string>;
+      }
     ) => BoundProfileProvider
   >
 >(
   (
-    configuration: {
-      services: IServiceSelector;
+    performResult: Result<unknown, ProfileParameterError | MapInterpreterError>,
+    configuration?: {
+      services?: IServiceSelector;
       profileProviderSettings?: NormalizedProfileProviderSettings;
-      security: SecurityConfiguration[];
+      security?: SecurityConfiguration[];
       parameters?: Record<string, string>;
-    },
-    performResult: Result<unknown, ProfileParameterError | MapInterpreterError>
+    }
   ) => ({
     ...Object.create(BoundProfileProvider.prototype),
-    configuration,
+    configuration: {
+      parameters: configuration?.parameters,
+      security: configuration?.security ?? [],
+      profileProviderSettings: configuration?.profileProviderSettings,
+      services: configuration?.services ?? new ServiceSelector([]),
+    },
     perform: jest.fn().mockResolvedValue(performResult),
   })
 );
